+++ NEWS.md
@@ -42,6 +42,8 @@
 - The `--relax=` analysis option is deprecated and is now equivalent to
   passing `--relaxed`.  The individual options are ignored.
 - Added support for generic subprograms in VHDL-2008.
+- New command `--install` allows easy installation of common third-party
+  packages such as OSVVM and UVVM.
 
 ## Version 1.6.2 - 2022-04-03
 - Fix `make -j` with GNU make (#440).